test: add before/after benchmarks
